--- src/main/java/io/zipcoder/casino/Game/Blackjack.java
@@ -1,17 +1,0 @@
-package io.zipcoder.casino.Game;
-
-import io.zipcoder.casino.card.Deck;
-import io.zipcoder.casino.utilities.Console;
-
-public class Blackjack {
-    Deck deck;
-    Console console;
-
-public Blackjack(Deck newDeck, Console newConsole) {
-    this.deck = newDeck;
-    this.console = newConsole;
-}
-
-
-
-}
