--- src/main/java/io/zipcoder/casino/Game/GoFish.java
@@ -1,4 +1,0 @@
-package io.zipcoder.casino.Game;
-
-public class GoFish {
-}
