+++ src/main/java/io/zipcoder/casino/game/Blackjack.java
@@ -1,0 +1,29 @@
+package io.zipcoder.casino.game;
+
+import io.zipcoder.casino.card.Deck;
+import io.zipcoder.casino.player.Player;
+import io.zipcoder.casino.utilities.Console;
+
+public class Blackjack {
+    private Deck deck;
+    private Console console;
+
+public Blackjack(Player newPlayer) {
+    super(newPlayer);
+    this.player =
+    this.deck = newDeck;
+    this.console = newConsole;
+}
+
+
+
+    public void Exit() {
+
+    }
+
+    public void play() {
+
+    }
+
+
+}
