+++ src/main/java/io/zipcoder/casino/game/GoFish.java
@@ -1,0 +1,4 @@
+package io.zipcoder.casino.game;
+
+public class GoFish {
+}
