--- src/test/java/io/zipcoder/casino/GameTest.java
@@ -1,20 +1,0 @@
-package io.zipcoder.casino;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class GameTest {
-
-    @Test
-    public void exit() {
-    }
-
-    @Test
-    public void play() {
-    }
-
-    @Test
-    public void playerTurn() {
-    }
-}
